Fix implementation of xadd on arm64
This went unnoticed for a long time because it only happens when the
cmpxchg loop runs multiple times. xadd is the only arithmetic operation
that changes the value of its source operand. This meant that if the
cmpxchg failed, the operation would be redone with the changed source
operand value instead of the original, giving the wrong result.