Fix handing for security-scoped URLs from picker
Apparently the API for these is a bit weird, first you call the
startAccessingSecurityScopedResource method and it returns whether the
URL is security scoped, and if so, it will give you access. In our case
we don't care, we just want access. Once you're done and the method
returns YES you are required to use stopAccessingSecurityScopedResource
to relinquish kernel resources, but the documentation states that this
is a no-op on anything you don't own so we can just call it on all URLs
unconditionally.